+++ src/app/feeds-frame/feeds.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders, HttpRequest } from '@angular/common/http';
 import { environment } from '../../environments/environment';
 import { PostActionForMedia } from 'models/post-action.model';
-import { Observable } from 'rxjs';
+import { Observable, of } from 'rxjs';
 import { User } from 'models/user.model';
 import { Page } from 'models/page.model';
 
@@ -16,6 +16,7 @@ export class FeedsService {
   constructor(private http: HttpClient) { }
 
   postFeed(data, apiUrl) {
+    if (!apiUrl) return of();
     const req = new HttpRequest('POST', this.baseUrl + apiUrl, data, { reportProgress: true });
     return this.http.request(req);
   }
@@ -26,30 +27,38 @@ export class FeedsService {
     return this.http.get<PostActionForMedia>(this.baseUrl + `user/posts/${postId}/media`);
   }
   getComments(postId, page) {
+    if (!postId) return of();
     return this.http.get(this.baseUrl + `user/posts/${postId}/comment`, { params: { page } });
   }
   postComment(postId, data) {
+    if (!postId) return of();
     return this.http.post(this.baseUrl + `user/posts/${postId}/comment`, data);
   }
   likePost(postId) {
+    if (!postId) return of();
     return this.http.post(this.baseUrl + `user/posts/${postId}/like`, {}, { observe: 'response' });
   }
   dislikePost(postId) {
+    if (!postId) return of();
     return this.http.delete(this.baseUrl + `user/posts/${postId}/like`, { observe: 'response' });
   }
   likeComment(commentId) {
+    if (!commentId) return of();
     return this.http.post(this.baseUrl + `user/posts/comment/${commentId}/like`, {});
   }
   dislikeComment(commentId) {
+    if (!commentId) return of();
     return this.http.delete(this.baseUrl + `user/posts/comment/${commentId}/like`);
   }
   getSharableLink(postId) {
+    if (!postId) return of();
     return this.http.post(this.baseUrl + `post/${postId}/link`, {});
   }
   removePost(postId) {
     const httpOptions = {
       headers: new HttpHeaders({ 'Content-Type': 'application/json' })
     };
+    if (!postId) return of();
     return this.http.delete(this.baseUrl + '/user/posts/' + postId, httpOptions);
   }
 }
